channelBox and chatBox socket -> Chat.tsx
