+++ frontend/src/components/chat/channelBox/ChannelBox.tsx
@@ -1,77 +1,26 @@
 import styles from './ChannelBox.module.css'
-import { Channel } from '../../../types/Channel'
 import CreateNewCh from './CreateNewCh'
 import ChannelList from './ChannelList'
-import { useEffect, useState } from 'react'
-import { io } from 'socket.io-client'
 import { CreateChannel } from '../../../types/CreateChannel'
-import { useAppDispatch, useAppSelector } from '../../../store/types'
-import { chatActions } from '../../../store/chat'
-import { UserData } from '../../../types/UserData'
+import { Channel } from '../../../types/Channel'
 
-const ChannelBox = () => {
-    const socket = io('http://localhost:8080')
-    const [allChan, setAllChan] = useState<Channel[]>([])
-    const userData = useAppSelector((state) => state.user.userData) as UserData
-    const dispatch = useAppDispatch()
+interface ChannelBoxProps {
+    allChan: Channel[]
+    handleCreation: (channel: CreateChannel) => void
+    deleteChannel: (channelId: number) => void
+    leaveChannel: (channelId: number) => void
+    joinChannel: (channelId: number, password: string) => void
+}
 
-    useEffect(() => {
-        getAllChannels()
-    }, [])
-
-    const createNewChannel = (channel: CreateChannel) => {
-        socket.emit('createNewChannel', channel, () => {})
-    }
-
-    socket.on('newChannel', () => {
-        getAllChannels()
-    })
-
-    const getAllChannels = () => {
-        socket.emit('getAllChannels', (response: any) => {
-            const allChannels = response
-            setAllChan(allChannels)
-        })
-    }
-
-    const handleCreation = (channel: CreateChannel) => {
-        createNewChannel(channel)
-    }
-
-    const LeaveChannel = (channelId: number) => {
-        socket.emit('leaveChannel', channelId, userData.user.id, () => {
-            dispatch(chatActions.selectChat(0))
-            getAllChannels()
-        })
-    }
-    const deleteChannel = (channelId: number) => {
-        socket.emit('deleteChannel', channelId, userData.user.id, () => {
-            dispatch(chatActions.selectChat(0))
-            getAllChannels()
-        })
-    }
-
-    const joinChannel = (channelId: number, password: string) => {
-        socket.emit(
-            'joinChannel',
-            channelId,
-            userData.user.id,
-            password,
-            () => {
-                getAllChannels()
-                dispatch(chatActions.selectChat(channelId))
-            }
-        )
-    }
-
+const ChannelBox = (props: ChannelBoxProps) => {
     return (
         <div className={styles.channelbox}>
-            <CreateNewCh handleCreation={handleCreation} />
+            <CreateNewCh handleCreation={props.handleCreation} />
             <ChannelList
-                allChan={allChan}
-                deleteChannel={deleteChannel}
-                leaveChannel={LeaveChannel}
-                joinChannel={joinChannel}
+                allChan={props.allChan}
+                deleteChannel={props.deleteChannel}
+                leaveChannel={props.leaveChannel}
+                joinChannel={props.joinChannel}
             ></ChannelList>
         </div>
     )
